+++ system/CogSolver/views.py
@@ -1,8 +1,5 @@
-from CogEditor.models import Application
-from CogSolver.forms import ApplicationForm
 from CogSolver.models import Rule, RuleEngine
 from django.shortcuts import render
-from django.views.generic.edit import CreateView
 
 
 def rules_report(request):
@@ -50,10 +47,3 @@ def rules_report(request):
     }
 
     return render(request, "CogSolver/rules_report.html", context)
-
-
-class ApplicationCreateView(CreateView):
-    model = Application
-    form_class = ApplicationForm
-    template_name = 'CogSolver/application_classifier.html'
-    success_url = '/solver/'
